clean single function with index option
