+++ halotools/mock_observables/counts_in_cells/__init__.py
@@ -1,3 +1,3 @@
 """
 """
-from .counts_in_cylinders import counts_in_cylinders, idx_in_cylinders
+from .counts_in_cylinders import counts_in_cylinders
