+++ halotools/mock_observables/counts_in_cells/engines/counts_in_cylinders_engine.pyx
@@ -261,6 +261,3 @@ def counts_in_cylinders_engine(
         return counts_uns, indexes1_uns, indexes2_uns
 
     return counts_uns
-
-# cdef append_to_indexes(idx1, idx2, indexes1, indexes2, n_idx):
-
